fix: Ensure chars are signed on all platforms
The default signedness of `char` is not defined in the C++ standard.
The SBPL code seems to assume that `char` is signed but on ARM platforms
it is unsigned, and thus SBPL is not working there.

The `-fsigned-char` flag in the CMakeLists.txt sets the default
signedness to "signed" independent of the platform, thus making it work
on ARM.